Move invites-panel.tsx into the invites route folder
Matches the co-location pattern used everywhere else in src/app/
(profile-form.tsx, welcome-form.tsx, programs-list.tsx, etc.) — the
file at the app root was a leftover from when the panel got extracted
out of the home page.
